Moved unit tests into folders corresponding to project folder location after refactor.
